refactor: solution 코드 정리
