+++ backend/src/main/java/develup/api/SolutionApi.java
@@ -3,6 +3,7 @@ package develup.api;
 import java.util.List;
 import develup.api.common.ApiResponse;
 import develup.domain.solution.SolutionRepository;
+import develup.domain.solution.SolutionSummary;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -17,7 +18,7 @@ class SolutionApi {
     }
 
     @GetMapping("/solutions")
-    public ResponseEntity<ApiResponse<List<SolutionSummaryResponse>>> getSolutions() {
+    public ResponseEntity<ApiResponse<List<SolutionSummary>>> getSolutions() {
         return ResponseEntity.ok(new ApiResponse<>(solutionRepository.findSummary()));
     }
 }
