+++ backend/src/main/java/develup/domain/solution/SolutionRepository.java
@@ -1,7 +1,6 @@
 package develup.domain.solution;
 
 import java.util.List;
-import develup.api.SolutionSummaryResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
@@ -10,10 +9,10 @@ public interface SolutionRepository extends JpaRepository<Solution, Long> {
     boolean existsByMember_IdAndMission_IdAndStatus(Long memberId, Long missionId, SolutionStatus status);
 
     @Query("""
-            select new develup.api.SolutionSummaryResponse(m.thumbnail, s.title, s.description)
-            from Solution s
-            join s.mission m
-            where s.status = 'COMPLETED'
+            SELECT new develup.domain.solution.SolutionSummary(s.id, m.thumbnail, s.title, s.description)
+            FROM Solution s
+            JOIN s.mission m
+            WHERE s.status = 'COMPLETED'
             """)
-    List<SolutionSummaryResponse> findSummary();
+    List<SolutionSummary> findSummary();
 }
